requirements.txt
# Core
pandas>=1.3
numpy>=1.21
scikit-learn>=1.0
matplotlib>=3.4
seaborn>=0.11

# Statistical analysis
scipy>=1.7
statsmodels>=0.13

# Working with dates and times
datetime

# Visualization enhancements
plotly>=5.0  

# Notebooks
jupyterlab 

# Data download
kaggle
gitpython